% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apde.chi.tools-package.R
\docType{package}
\name{apde.chi.tools-package}
\alias{apde.chi.tools}
\alias{apde.chi.tools-package}
\title{apde.chi.tools: Simplify Community Health Indicators Analytics Pipeline for Public Health Seattle & King County's Assessment Policy Development and Evaluation Unit}
\description{
This package provides tools for epidemiologists within Public Health Seattle
and King County. It aligns with analytics guidelines and practices for Community
Health Indicators (CHI) as reported by the Assessment, Policy Development, and
Evaluation Unit. The package simplifies the execution of identical CHI analyses
across multiple variables, creates CHI Tableau-ready output of the analysis,
performs quality assurance regarding data structure and formatting, and
facilitates pushing and pulling data to APDE servers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PHSKC-APDE/apde.chi.tools}
  \item Report bugs at \url{https://github.com/PHSKC-APDE/apde.chi.tools/issues}
}

}
\author{
\strong{Maintainer}: Ronald Buie \email{rbuie@kingcounty.gov}

Authors:
\itemize{
  \item Ronald Buie \email{rbuie@kingcounty.gov}
}

}
\keyword{internal}
