% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_chi_count_by_age.R
\name{chi_count_by_age}
\alias{chi_count_by_age}
\title{CHI Count By Age}
\usage{
chi_count_by_age(ph.data = NULL, ph.instructions = NULL, source_date = NULL)
}
\arguments{
\item{source_date}{}
}
\description{
CHI Count By Age
}
