% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_generate_tro_shell.R
\name{chi_generate_tro_shell}
\alias{chi_generate_tro_shell}
\title{CHI generate TRO shell}
\usage{
chi_generate_tro_shell(
  ph.analysis_set,
  start.year,
  end.year,
  year.span = NULL,
  trend.span = NULL,
  trend.periods = NULL
)
}
\arguments{
\item{ph.analysis_set}{name of data.table to parse}

\item{start.year}{the earliest year to be used for estimates}

\item{end.year}{the latest year to be used for aggregate estimates (note, the earliest year for trends estimates is calculated from from the span and number of periods)}

\item{year.span}{the number of years to be included in a single non-trend period}

\item{trend.span}{the number of years to be included in a single trend period}

\item{trend.periods}{the number of periods to be included in a trend}
}
\value{
data table with a single row for each calculation to be performed in generating Tableau Ready Output for CHI reporting
}
\description{
This function takes an analysis set file and indicator of which set should be processed. It returns a skeleton of CHI Tableau Ready Output.
}
\details{
It takes in a data.table containing a compact list of variables, byvariables, and analysis types, and returns a shell table of the rows and columns expected in a CHI Tableau ready output. For details on TRO format, review here: https://kc1.sharepoint.com/:x:/r/teams/DPH-CommunityHealthIndicators/CHIVizes/CHI-Standards-TableauReady%20Output.xlsx?d=wbed2f507b8344d288658c5724f64c001&csf=1&web=1&e=qEIPcc&nav=MTVfezAwMDAwMDAwLTAwMDEtMDAwMC0wMjAwLTAwMDAwMDAwMDAwMH0

the expected format of the analysis file is:
set: numeric integer 1...x, indicates set the observations are calcualted as part of (why are sets valueable? should this be discarded?)
cat1: character, the name expected in CHI TRO for cat1
cat1_varname: character, the name expected in CHI TRO for cat1_varname
kingcounty: character "":"X", indicator of if analysis is king county specific (could be removed, this is imputable by variable name)
wastate: character "":"x", indicator of if analysis is of wa state
demgroups: character "":"x", indicator of if analysis includes single demographic
crosstabs: character "":"x", indicator of if analysis includes crosstabulations
trands: character "":"x", indicator of if analysis includes trends
set_idictaor_keys character comma sep list, list of indicators variables expected from data source
}
\keyword{CHI,}
\keyword{Production}
\keyword{Tableau,}
