% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_process_nontrends.R
\name{chi_process_nontrends}
\alias{chi_process_nontrends}
\title{CHI generate nontrend years}
\usage{
chi_process_nontrends(ph.analysis_set = NULL, myset = NULL)
}
\arguments{
\item{ph.analysis_set}{name of a formal \code{analysis_set} as described in the
\strong{Details} section below}

\item{myset}{chosen set number from table}
}
\value{
Returns a data table of calculation instructions with columns:
\itemize{
  \item \code{indicator_key} Unique identifier for each indicator
  \item \code{tab} Analysis type (_kingcounty, demgroups, etc.)
  \item \code{cat1} A standard value from the \code{'cat'} column of \code{\link[rads.data]{misc_chi_byvars}}
  \item \code{cat1_varname} A standard value from the \code{'varname'} column of \code{\link[rads.data]{misc_chi_byvars}}
  \item \code{cat2} A standard value from the \code{'cat'} column of \code{\link[rads.data]{misc_chi_byvars}}
  \item \code{cat2_varname} A standard value from the \code{'varname'} column of \code{\link[rads.data]{misc_chi_byvars}}
  \item \code{start} Starting year of the time period to be analyzed
  \item \code{end} Ending year of the time period to be analyzed
}
}
\description{
\strong{Internal}, used by \code{\link{chi_generate_tro_shell}}
}
\details{
This function takes an analysis set table created by \code{\link{chi_generate_analysis_set}}
and returns analysis instructions for non-trend analyses.

The expected format of \code{ph.analysis_set} is:
  \itemize{
    \item \code{cat1}: character. A standard value from the \code{'cat'} column
    of \code{\link[rads.data]{misc_chi_byvars}}, e.g. \code{'Gender'}
    \item \code{cat1_varname}: character. A standard value from the \code{'varname'}
    column of \code{\link[rads.data]{misc_chi_byvars}}, e.g., \code{'chi_sex'}
    \item \code{_kingcounty}: character \code{'x'}. Is the analysis for King County as a whole?
    \item \code{_wastate}: character \code{'x'}. Is the analysis is for WA state as a whole?
    \item \code{demgroups}: character \code{'x'}. Is the analysis for demgroups?
    \item \code{crosstabs}: character \code{'x'}. Is the analysis for crosstabs?
    \item \code{trends}: character \code{'x'}, Is the the analysis for trends?
    \item \code{set}: Integer identifying groups of indicators with identical analysis patterns
    \item \code{set_indicator_keys}: Comma-separated list of indicator keys sharing the pattern
  }
}
\seealso{
\code{\link{chi_generate_tro_shell}}, which calls on this function to create
ph.instructions

\code{\link{chi_process_trends}}, which is a counterpart to this function and
is also an internal function used by \code{\link{chi_generate_analysis_set}}
}
\keyword{CHI,}
\keyword{Production,}
\keyword{Tableau,}
\keyword{internal}
