% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_process_trends.R
\name{chi_process_trends}
\alias{chi_process_trends}
\title{CHI Generate Trend Years}
\usage{
chi_process_trends(
  indicator_key = NULL,
  trend.span = NULL,
  end.year = NULL,
  trend.periods = NULL
)
}
\arguments{
\item{indicator_key}{chi indicator key variable}

\item{trend.span}{the number of years to be included in a single trend period}

\item{end.year}{last year of a trend year time series}

\item{trend.periods}{number of periods to calculate}
}
\value{
Returns a data table of trend calculation instructions with columns:
\itemize{
  \item \code{indicator_key} Unique identifier for each indicator
  \item \code{tab} Always \code{'trends'}
  \item \code{cat1} A standard value from the \code{'cat'} column of \code{\link[rads.data]{misc_chi_byvars}}
  \item \code{cat1_varname} A standard value from the \code{'varname'} column of \code{\link[rads.data]{misc_chi_byvars}}
  \item \code{cat2} A standard value from the \code{'cat'} column of \code{\link[rads.data]{misc_chi_byvars}}
  \item \code{cat2_varname} A standard value from the \code{'varname'} column of \code{\link[rads.data]{misc_chi_byvars}}
  \item \code{start} Starting year of the time period to be analyzed
  \item \code{end} Ending year of the time period to be analyzed
}
}
\description{
\strong{Internal}, used by \code{\link{chi_generate_tro_shell}}
}
\details{
This function is used internally by \code{\link{chi_generate_analysis_set}}
to create instructions for trend analyses.
}
\seealso{
\code{\link{chi_generate_tro_shell}}, which calls on this function to create
ph.instructions

\code{\link{chi_process_nontrends}}, which is a counterpart to this function and
is also an internal function used by \code{\link{chi_generate_analysis_set}}
}
\keyword{CHI,}
\keyword{Production,}
\keyword{Tableau,}
\keyword{internal}
