% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_sql_update.R
\name{chi_sql_update}
\alias{chi_sql_update}
\title{CHI SQL Update}
\usage{
chi_sql_update(
  CHIestimates = NULL,
  CHImetadata = NULL,
  table_name = NULL,
  server = "development",
  replace_table = F
)
}
\arguments{
\item{CHIestimates}{DT or DF containing CHI analytic results}

\item{CHImetadata}{DT or DF containing CHI metadata}

\item{table_name}{name of SQL Server table to update}

\item{server}{type of server ('development' for KCITSQLUATHIP40 and 'production' for KCITSQLPRPHIP40)}

\item{replace_table}{If T, drop existing table and insert data, if F update matching rows and insert new data}
}
\value{
status message indicating success and location, or failure, of upload
}
\description{
function to update (or replace) results and metadata into APDE PHExtractStore servers. For details on server access, and to configure your local settings please review documentation [here](https://kc1.sharepoint.com/:w:/r/teams/DPH-KCCross-SectorData/_layouts/15/Doc.aspx?sourcedoc=%7B34352D66-9CD6-45C9-AD19-B8FE88A4C7C6%7D&file=SQL%20Server%20Setup%20APDE.docx&action=default&mobileredirect=true)
}
\keyword{CHI,}
\keyword{Production}
\keyword{Tableau,}
