% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{chi_standard_varnames}
\alias{chi_standard_varnames}
\title{Deprecated: \code{chi_standard_varnames} has moved to apde.etl}
\usage{
chi_standard_varnames
}
\description{
\code{chi_standard_varnames} has been moved to \code{apde.etl::chi_standard_varnames}.
}
\details{
For the time being, \code{apde.chi.tools::chi_standard_varnames} silently returns
\code{apde.etl::chi_standard_varnames} via an active binding \code{R/zzz.R}.
}
\keyword{internal}
