% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_update_sql.R
\name{chi_update_sql}
\alias{chi_update_sql}
\title{CHI SQL Update}
\usage{
chi_update_sql(
  CHIestimates = NULL,
  CHImetadata = NULL,
  table_name = NULL,
  server = "development",
  replace_table = FALSE
)
}
\arguments{
\item{CHIestimates}{data.frame/data.table containing CHI analytic results}

\item{CHImetadata}{data.frame/data.table containing CHI metadata}

\item{table_name}{name of SQL Server table to update, e.g., \code{'acs'}, \code{'birth'}, \code{'death'}, etc.}

\item{server}{type of server (\code{'development'} for KCITSQLUATHIP40 and
\code{'production'} for KCITSQLPRPHIP40)

Default \code{server = 'development'}}

\item{replace_table}{If \code{TRUE}, the existing SQL table is dropped and fully replaced with the
new data. If \code{FALSE}, rows matching the \code{indicator_key} values in \code{CHIestimates} and
\code{CHIestimates} are deleted from SQL and then replaced with the new data, leaving all other rows
intact.

Default \code{replace_table = FALSE}}
}
\value{
Returns a status message indicating success and location, or failure, of upload
}
\description{
Function to update (or replace) results and metadata in APDE PHExtractStore
servers.
}
\details{
For details on server access, and to configure your local settings,
please review documentation:
SharePoint > DPH-KCCross-SectorData > Documents > References > SQL > SQL Server Setup APDE.docx.
}
\examples{
\dontrun{
# Update development database
chi_update_sql(
  CHIestimates = final_estimates,
  CHImetadata = final_metadata,
  table_name = "birth",
  server = "development",
  replace_table = FALSE
)
}

}
\seealso{
\code{\link{chi_qa_tro}} for validating data before upload

\code{\link{chi_compare_estimates}} for comparing with existing data
}
\keyword{CHI,}
\keyword{Production}
\keyword{Tableau,}
