% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_chi_calc.R
\name{chi_calc}
\alias{chi_calc}
\title{Title}
\usage{
chi_calc(
  ph.data = NULL,
  ph.instructions = NULL,
  rate = F,
  rate_per = NULL,
  small_num_suppress = T,
  suppress_low = 0,
  suppress_high = 9,
  source_name = "blahblah",
  source_date = NULL
)
}
\arguments{
\item{source_date}{}
}
\description{
Title
}
