% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proto_chi_drop_illogical_ages.R
\name{chi_drop_illogical_ages}
\alias{chi_drop_illogical_ages}
\title{Drop Illogical Age Combinations from CHI Data}
\usage{
chi_drop_illogical_ages(ph.data, agevar = "chi_age")
}
\arguments{
\item{ph.data}{Input data.table to process}

\item{agevar}{Name of the age variable (defaults to 'chi_age')}
}
\value{
A data.table with illogical age combinations removed
}
\description{
Removes age combinations that don't make logical sense based on category
groupings. For example, removes cases where the age falls outside the range
specified by age group categories.
}
