% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_chi_drop_illogical_ages.R
\name{chi_drop_illogical_ages}
\alias{chi_drop_illogical_ages}
\title{CHI_drop_illogical_ages}
\usage{
chi_drop_illogical_ages(DTx, agevar = "chi_age")
}
\arguments{
\item{agevar}{}
}
\description{
a helper function for filtering ages from a set of observations
}
\details{
CHI analyses relay
}
