% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_chi_generate_instructions_pop.R
\name{chi_generate_instructions_pop}
\alias{chi_generate_instructions_pop}
\title{Title}
\usage{
chi_generate_instructions_pop(mycount.data, povgeo = NA)
}
\arguments{
\item{povgeo}{}
}
\description{
Title
}
