% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_generate_metadata.R
\name{chi_generate_metadata}
\alias{chi_generate_metadata}
\title{CHI Generate Metadata}
\usage{
chi_generate_metadata(meta.old = NULL, est.current = NULL)
}
\arguments{
\item{meta.old}{Previous metadata table}

\item{est.current}{current year's tableau ready output with completed estimates}
}
\value{
table of metadata
}
\description{
function to generate metadata table combining existing metadata and latest estimates.
}
