% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_process_trends.R
\name{chi_process_trends}
\alias{chi_process_trends}
\title{CHI Generate Trend Years}
\usage{
chi_process_trends(
  indicator_key = NULL,
  trend.span = NULL,
  end.year = NULL,
  trend.periods = NULL
)
}
\arguments{
\item{indicator_key}{chi indicator key variable}

\item{trend.span}{the number of years to be included in a single trend period}

\item{end.year}{last year of a trend year time series}

\item{trend.periods}{number of periods to calculate}
}
\value{
TRO with rows for each indicator key and span of years within the provided time frame
}
\description{
helper function for chi_generate_tro_shell
}
\details{
called by chi_generate_tro_shell to calculate and create rows for expected trends analyses.
}
\keyword{CHI,}
\keyword{Production,}
\keyword{Tableau,}
\keyword{internal}
