% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_qa_tro.R
\name{chi_qa_tro}
\alias{chi_qa_tro}
\title{CHI - QA: Tableau Ready output}
\usage{
chi_qa_tro(
  CHIestimates,
  CHImetadata,
  acs = FALSE,
  ignore_trends = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{CHIestimates}{Name of a data.table or data.frame containing the prepared data to be pushed to SQL}

\item{CHImetadata}{Name of a data.table or data.frame containing the metadata to be pushed to SQL}

\item{acs}{default FALSE, Indicates whether it is ACS data (which does not have / need varnames)}

\item{ignore_trends}{default TRUE, Indicates whether the time_trends column should be ignored when checking for missing data.}

\item{verbose}{default TRUE, if false will only return status}
}
\value{
1 or 0 for pass or fail status
}
\description{
Checks estimates created by \code{\link{chi_calc}} and metadata created by
\code{\link{chi_generate_metadata}} for compliance with CHI TRO standard.
}
\details{
This function tests if the structure of the data matches CHI Tableau Ready Output
specifications. If set to verbose mode (default) will report diagnostic information in
The user interface as warnings (if failed) and messages (for progress and pass). In any
Case will return 1 or 0 for pass or fail respectively.
The CHI Tableau Ready Output (TRO) standards can be reviewed here:
\href{https://kc1.sharepoint.com/teams/DPH-CommunityHealthIndicators/CHIVizes/CHI-Standards-TableauReady\%20Output.xlsx}{
SharePoint > Community Health Indicators > CHI_vizes > CHI-Standards-TableauReady Output.xlsx}
}
\examples{

\dontrun{
# QA check estimates and metadata
qa_status <- chi_qa_tro(
  CHIestimates = my_estimates,
  CHImetadata = my_metadata,
  acs = FALSE,
  ignore_trends = TRUE,
  verbose = TRUE
)
}

}
\seealso{
\code{\link{chi_calc}} for generating estimates

\code{\link{chi_generate_metadata}} for creating metadata

\code{\link{chi_compare_estimates}} for comparing estimates between versions
}
\keyword{CHI,}
\keyword{Production}
\keyword{Tableau,}
