% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_qa_tro.R
\name{chi_qa_tro}
\alias{chi_qa_tro}
\title{CHI - QA: Tableau Ready output}
\usage{
chi_qa_tro(chi_est, chi_meta, acs = F, ignore_trends = T, verbose = F)
}
\arguments{
\item{chi_est}{Name of a data.table or data.frame containing the prepared data to be pushed to SQL}

\item{chi_meta}{Name of a data.table or data.frame containing the metadata to be pushed to SQL}

\item{acs}{default FALSE, Indicates whether it is ACS data (which does not have / need varnames)}

\item{ignore_trends}{default TRUE, Indicates whether the time_trends column should be ignored when checking for missing data.}

\item{verbose}{default TRUE, if false will only return status}
}
\value{
1 or 0 for pass or fail status
}
\description{
Checks data for compliance with CHI TRO standard
}
\details{
This function tests if the structure of the data matches CHI Tableau Ready Output
specifications. If set to verbose mode (default) will report diagnostic information in
The user interface as warnings (if failed) and messages (for progress and pass). In any
Case will return 1 or 0 for pass or fail respectively.
The CHI Tableau Ready Output (TRO) standards can be reviewed [here]("//phshare01/epe_share/WORK/CHI Visualizations/Tableau Ready Output Format_v2.xlsx")
}
\examples{

\dontrun{
# create sample data

# run function
}

}
\keyword{CHI,}
\keyword{Production}
\keyword{Tableau,}
