% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_sql_update.R
\name{chi_sql_update}
\alias{chi_sql_update}
\title{CHI SQL Update}
\usage{
chi_sql_update(
  CHIestimates = NULL,
  CHImetadata = NULL,
  table_name = NULL,
  server = "wip",
  replace_table = F
)
}
\arguments{
\item{CHIestimates}{DT or DF containing CHI analytic results}

\item{CHImetadata}{DT or DF containing CHI metadata}

\item{table_name}{name of SQL Server table to update}

\item{server}{name of serverto access}

\item{replace_table}{If T, drop existing table and insert data, if F update matching rows and insert new data}
}
\value{
status message indicating success and location, or failure, of upload
}
\description{
function to update (or replace) results and metadata in SQL 51 (dev/WIP) or 50 (prod)
}
